Change to float the argpase tversky values

+++ Common/Unet_Medical_Parse.py
@@ -84,9 +84,9 @@ def create_arg_parser():
 
     ### Loss control
     parser.add_argument('--loss', type=str, default='WBCE_DiceLoss', help='Which loss used')
-    parser.add_argument('--WBCE-diceloss', type=int, default=0.5, help='Weighted Binary Cross Entropy coefficient')
-    parser.add_argument('--tversky-alpha', type=int, default=0.3, help='Tversky Loss alpha coefficient')
-    parser.add_argument('--tversky-beta', type=int, default=0.7, help='Tversky Loss beta coefficient')
+    parser.add_argument('--WBCE-diceloss', type=float, default=0.5, help='Weighted Binary Cross Entropy coefficient')
+    parser.add_argument('--tversky-alpha', type=float, default=0.3, help='Tversky Loss alpha coefficient')
+    parser.add_argument('--tversky-beta', type=float, default=0.7, help='Tversky Loss beta coefficient')
 
     # parser.add_argument('--data-split', choices=['val', 'test'], required=True,
     #                     help='Which data partition to run on: "val" or "test"')
